Create `GuiSettings` helper class
